+++ changelog.md
@@ -13,6 +13,7 @@
 - Improved the #results message to link to thre game thread when possible
 - LFG Posts can now be cancelled by guild moderators
 - Clicking a selected time in Compare Availability deselects that time block
+- Markdown added to Series and Survey descriptions
 
 ### Bug Fixes
 - Subscribing to a scheduling poll now notifies on every new "Yes" answer instead of just poll close.